Change Cy0 back to 0.0
With the value on 0.4, the plane was pitching up a lot, then stalling,
then pitching down, and then when it gains speed pitching up again.

It seems like this value should be 0.0, based on the following comment
from the beginners guide:
"Cy0 will almost always be 0. It is the angle of attack where the wing
produces zero lift. For symmetric airfoils, this is always 0. Larger
aircraft are likely to use a wing with camber and this value will be
non-zero in radians."